moved footer pages to pages folder
